refactor: simplifie URL returnPath parameter handling
PT-1886
